+++ inscripta/biocantor/io/genbank/exc.py
@@ -27,3 +27,11 @@ class GenBankExportError(BioCantorIOException):
     """
 
     pass
+
+
+class GenBankLocationException(GenBankParserError):
+    """
+    Raised when there is an issue with locations (usually BioPython makes them None)
+    """
+
+    pass