+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.4]
+### Fixed
+- Handle genbank files with broken intervals gracefully.
+- Fix interval parsing for negative strand features.
+
+### Changed
+- The tag `Name` can now be used to identify a feature interval in a GFF3/GenBank file.
+
 
 ## [0.4.3]
 ### Fixed